Add configuration options for drip rate
In the previous configuration, `drip_interval` was used to scale
the store volumes (`kststor1` etc.) into drip rates.  Now,
`driprate_store_empty` specifes the driprate that the model is expected
to produce when the store is empty and `driprate_store_full` is for when
the store is full, with linear interpolation in between.